Added api routes as per corrected understanding
